Connect plumbing to search for manifest files recursively
Check for "-r" or "--recursive" parameter. If found, use the MANIFEST_PATH
as root to recursively search for manifest files. Each valid manifest is
added to a list of manifest files and will be treated individually. For
now, valid manifest files are considered to use .txt or .csv extensions.

Fix coding style somehow.

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>